# test: developed a script to show summary of cargo test