Developed a script to show summary of cargo test in brief which we using to compare variances of failed tests cases.
- It running `cargo test` by Rust code which make command faster and avoid any output causing delay.
- In case there are failure cases, it sorts them, so we can compare with previous tests just by opening the log file.
- Old command `./scripts/test` can be used.
- Initial test between old command `./scripts/test` and new command `./scripts/test summary` is around a minute as below output feedback, but after adding some functions to the new script it is the same time, but with a feature of summarized data.

Old command to run tests with complete output:
```shell
❯ time ./scripts/test 
   Compiling zinzen v0.1.1 (/home/moaz/dev/entity/tijl/ZinZen-scheduler)
    Finished test [unoptimized + debuginfo] target(s) in 4.17s
     Running unittests src/lib.rs (target/debug/deps/scheduler-fef0136fcda33c72)
     Running unittests src/bin/flamegraph-bin.rs (target/debug/deps/scheduler_bin-fe7e9f612de3579f)
     Running tests/rust_tests.rs (target/debug/deps/rust_tests-e6c3ee9fa4fcce89)
error: test failed, to rerun pass `--test rust_tests`
   Doc-tests scheduler
error: 1 target failed:
    `--test rust_tests`

real	9m48.163s
user	16m23.878s
sys	0m5.636s
```

New command without any output or filtering:
```shell
❯ time cargo run --bin tester
   Compiling zinzen v0.1.1 (/home/moaz/dev/entity/tijl/ZinZen-scheduler)
    Finished dev [unoptimized + debuginfo] target(s) in 2.25s
     Running `target/debug/tester`

=======
=======


real	8m31.571s
user	15m7.376s
sys	0m4.973s
```
New command after finalized which have filter for output and sorting:
```shell
❯ time ./scripts/test summary
   Compiling zinzen v0.1.1 (/home/moaz/dev/entity/tijl/ZinZen-scheduler)
    Finished dev [unoptimized + debuginfo] target(s) in 2.19s
     Running `target/debug/test-summary`


real	8m8.459s
user	14m33.324s
sys	0m3.982s
```

- Script used by calling `./scripts/test summary`,
- Output will be saved to files starting with formatting `summary_<DATE_TIME>.log` like `summary_2023-07-13--19-27.log`
- Below is output of the script in the file
```shell
failures:

    e2e::children_with_over_duration
    e2e::demo_2_with_filler_with_budget
    e2e::different_repetition_goals
    e2e::every_6_hours
    e2e::every_6_hours_2
    e2e::flex_duration_2
    e2e::goals_dependency
    e2e::goals_hierarchy
    e2e::i293_postpone_2
    e2e::i309
    e2e::impossible_2
    e2e::issue_276_weekdays_filter_on_budget
    e2e::issue_284_filter_days_of_week_with_impossible
    e2e::issue_284_no_repeat_large_dur_with_impossible
    e2e::planned_goals_hierarchy
    e2e::show_sleep
    e2e::slot_reduced_1


test result: FAILED. 49 passed; 17 failed; 0 ignored; 0 measured; 0 filtered out; finished in 482.87s

```